Untrack test binaries and fix dump_beast_check 2038 overflow
tracetest and cprtests are Makefile targets removed by `make clean`, so
they churn on every build and conflict on merge. Their debug info also
embedded local build paths. Untrack them and gitignore the test binaries
alongside the already-ignored readsb/viewadsb.

dump_beast_check computed its file index into an int32_t, which overflows
in 2038 when dump_interval is 1 (index is then epoch-seconds). The
(time_t) cast added in f332158 would sign-extend the wrong value.
Widen both the local and Modes.dump_beast_index to int64_t.

+++ readsb.h
@@ -869,7 +869,7 @@ struct _Modes
     zstd_fw_t *dump_fw;
     int64_t dump_next_ts; // last timestamp sent
     int32_t dump_interval;
-    int32_t dump_beast_index;
+    int64_t dump_beast_index;
     uint64_t dump_lastReceiverId;
     int8_t dump_compressionLevel;
     int8_t beast_dated_dirs; // --devel=beast_dated_dirs: organise captures into YYYY-MM-DD subdirs
